[DASH-27] - correct class names
