+++ integration-tests/src/test/java/com/vaadin/addon/board/testbenchtests/Dash27HorizontalLayoutUIIT.java
@@ -2,7 +2,7 @@ package com.vaadin.addon.board.testbenchtests;
 
 import org.junit.Test;
 
-import com.vaadin.addon.board.testUI.Dash27HorizontalLayoutUI;
+import com.vaadin.addon.board.testUI.Dash27LayoutComponents;
 
 /**
  *
@@ -10,7 +10,7 @@ import com.vaadin.addon.board.testUI.Dash27HorizontalLayoutUI;
 public class Dash27HorizontalLayoutUIIT extends AbstractParallelTest {
     @Override
     protected Class<?> getUIClass() {
-        return Dash27HorizontalLayoutUI.class;
+        return Dash27LayoutComponents.HorizontalLayoutUI.class;
     }
 
     @Test
